Change SceneTemplate to mostly use a local
There are I think two benefits to doing it this way. First when copying
the template into your own code, you only have to change the name in 3
places vs doing a find and replace. And secondly, after you have a scene
or two in your own code, it makes it a lot easier to copy and paste the
function declaration and call to super from one scene to another with no
changes at all.