Defer Noble.transition to update loop
Before it was possible to call transition at times when it was not safe
to do so, such as in input handlers. It would sometimes work, but then
sometimes crash the playdate. It was possible to call a transition
multiple times resulting in a bonk.

So instead, make the transition function enqueue the transition until
the next game loop, and also take care of making sure that other calls
are ignored until the transition is completed.

This should allow transition to be called anywhere, even during button
presses that otherwise would require a lot of boilerplate flag setting
and checking.t