Fix uneven animation frame durations in Noble.Animation
This changelist addresses an issue in the `Noble.Animation` module where the first frame of an animation was displayed for a longer duration than subsequent frames.

The problem was due to the way `previousFrameDurationCount` was reset on successive loops or transitions. This has been corrected to ensure all frames have the same duration.