Add an optional table argument which is passed to scene's init
When calling Noble.transition (and Noble.new) it is sometimes a good idea to pass some arguments to initialize the scene. For example, we could have a `Fight` scene that we want to initialize with `{Player1 = "Ryu", Player2 = "Ken", Arena = "Airport"} to signify that Ryu will be fighting Ken in the Airport.

This is a nice encapsulation and allows the developer to avoid having global variables to pass as arguments.

This pull request is similar to https://github.com/NobleRobot/NobleEngine/pull/61 but I think it improves it as it doesn't use `...`. This is a key difference since it allows for new arguments to be set for both Noble.transition and Noble.new in the future (if needed).

This change is retrocompatible and users who pull the repo won't be forced to make changes as:
A) Users which use the previous way of calling Noble.transition and Noble.new will have a `nil` parameter appended to the call which is handled correctly, and
B) Users who defined their init functions without the extra argument will automatically ignore the `sceneArgs` even if it's passed.